pgml-apps/rag-retrieval-timing-tests/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
Authlib==1.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
distro==1.9.0
grpcio==1.62.0
grpcio-health-checking==1.62.0
grpcio-tools==1.62.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx==0.27.0
hyperframe==6.0.1
idna==3.6
numpy==1.26.4
openai==1.13.3
pgml==1.0.0
pinecone-client==3.1.0
portalocker==2.8.2
protobuf==4.25.3
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
qdrant-client==1.7.3
requests==2.31.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
validators==0.22.0
weaviate-client==4.5.1
